Renamed resend to resend, made teststeps less technical, renamed classes to be in line with existing test step implementation classes.
