+++ cucumber-tests-platform-smartmetering/src/test/resources/features/osgp-adapter-ws-smartmetering/notifications/ResendNotifications.feature
@@ -1,4 +1,4 @@
-@SmartMetering @Platform @test
+@SmartMetering @Platform
 Feature: SmartMetering notifications - Resend notifications
   As an OSGP user
   I want the platform to resend missed notifications
@@ -11,9 +11,9 @@ Feature: SmartMetering notifications - Resend notifications
       | MessageType               | REQUEST_PERIODIC_METER_DATA                            |
       | CorrelationUid            | test-org\|\|\|TEST1024000000001\|\|\|20170101000000000 |
       | NumberOfNotificationsSent |                                                      0 |
-    When the missed notification is resend
+    When the missed notification is resent
       | CorrelationUid | test-org\|\|\|TEST1024000000001\|\|\|20170101000000000 |
-    Then a record in the response_data table of the database has values
+    Then the response data has values
       | DeviceIdentification      | TEST1024000000001                                      |
       | MessageType               | REQUEST_PERIODIC_METER_DATA                            |
       | CorrelationUid            | test-org\|\|\|TEST1024000000001\|\|\|20170101000000000 |
@@ -25,8 +25,8 @@ Feature: SmartMetering notifications - Resend notifications
       | MessageType               | REQUEST_PERIODIC_METER_DATA                            |
       | CorrelationUid            | test-org\|\|\|TEST1024000000001\|\|\|20170101000000000 |
       | NumberOfNotificationsSent |                                                      0 |
-    When no notification is resend
-    Then a record in the response_data table of the database has values
+    When no notification is resent
+    Then the response data has values
       | DeviceIdentification      | TEST1024000000001                                      |
       | MessageType               | REQUEST_PERIODIC_METER_DATA                            |
       | CorrelationUid            | test-org\|\|\|TEST1024000000001\|\|\|20170101000000000 |
@@ -39,8 +39,8 @@ Feature: SmartMetering notifications - Resend notifications
       | MessageType               | REQUEST_PERIODIC_METER_DATA                            |
       | CorrelationUid            | test-org\|\|\|TEST1024000000001\|\|\|20170101000000000 |
       | NumberOfNotificationsSent |                                                      3 |
-    When no notification is resend
-    Then a record in the response_data table of the database has values
+    When no notification is resent
+    Then the response data has values
       | DeviceIdentification      | TEST1024000000001                                      |
       | MessageType               | REQUEST_PERIODIC_METER_DATA                            |
       | CorrelationUid            | test-org\|\|\|TEST1024000000001\|\|\|20170101000000000 |
